motor parameter conversion has been verified, and rough adjustment r/min to m/s parameters.

+++ src/open_loop_motor_control/include/open_loop_motor_control/movement_control.hpp
@@ -30,7 +30,7 @@
 #define TOPIC_cmd_vel "cmd_vel"
 #define TOPIC_odom "odom"
 #define queue_size 100
-#define running_cycle 0.1
+#define running_cycle (double)0.1
 
 /* Define End */
